1. 优化IM Demo联系人列表加载速度 2. 在VS2012下编译通过 3. 增加一个动画2工程 4. 优化动画创建接口，修改RE,FLASH控件刷新方式等其它bug

+++ trunk/UI/UISDK/Kernel/Src/SkinParse/datasource/File/filebufferreader.cpp
@@ -17,7 +17,7 @@ void  FileBufferReader::Release()
 
 int   FileBufferReader::read(char* pread, int nread)
 {
-    return f.read(pread, nread).gcount();
+    return (int)f.read(pread, nread).gcount();
 }
 void  FileBufferReader::seek(int npos, ios_base::seek_dir dir)
 {
@@ -25,7 +25,7 @@ void  FileBufferReader::seek(int npos, ios_base::seek_dir dir)
 }
 int  FileBufferReader::tell()
 {
-    return f.tellg();
+    return (int)f.tellg();
 }
 bool  FileBufferReader::load(const TCHAR* szPath)
 {
